[Relax] Default purity as unknown, instead of pure
Prior to this commit, the constructors for `relax.Function` and
`relax.FuncStructInfo` provided a default annotation as a pure
function.  While this could be overridden, it

This commit updates the IR to store purity as `Optional<Bool>` instead
of `bool`, and changes the default to `NullOpt`.

A function's purity can have three possible values:

- `Bool(true)`: The function is known to be pure.
- `Bool(false)`: The function is known to be impure.
- `NullOpt`: The function's purity is unknown.

In most cases `Bool(false)` and `NullOpt` should be treated
equivalently, as they both indicate a function that may contain side
effects.  However, in the future, inference of purity may occur when
the purity is `NullOpt`, based on analysis of a function's body.  A
function with purity of `Bool(false)` is known to be impure, and
further analysis is unnecessary.

+++ tests/python/relax/distributed/test_distributed_dtensor_sinfo.py
@@ -19,8 +19,9 @@ import tvm
 import tvm.testing
 import pytest
 
-from tvm import relax as rx, TVMError, tir
-from tvm.ir import structural_equal, Range
+from tvm import relax as rx, tir
+from tvm.ir import Range
+from tvm.script import tir as T
 
 
 def _check_equal(x, y, map_free_vars=False):
@@ -66,7 +67,7 @@ def test_dtensor_struct_info():
     assert s0 == s1
     tvm.ir.assert_structural_equal(s0.device_mesh, device_mesh0)
     assert s0.device_mesh.shape == (2, 2)
-    tvm.ir.assert_structural_equal(s0.device_mesh.device_range, Range(0, 4))
+    tvm.ir.assert_structural_equal(s0.device_mesh.device_range, Range(T.int64(0), T.int64(4)))
     tvm.ir.assert_structural_equal(s0.placement, placement0)
     assert len(s0.placement.dim_specs) == 2
     assert s0.placement.dim_specs[0] == shard0
